+++ src/LiveComponent/assets/src/Component/plugins/LoadingPlugin.ts
@@ -2,7 +2,7 @@ import {
     Directive,
     DirectiveModifier,
     parseDirectives
-} from '../../directives_parser';
+} from '../../Directive/directives_parser';
 import { combineSpacedArray}  from '../../string_utils';
 import BackendRequest from '../../BackendRequest';
 import Component from '../../Component';