+++ src/LiveComponent/assets/src/HookManager.ts
@@ -5,14 +5,21 @@ export default class {
         this.hooks = new Map();
     }
 
-    /**
-     * Add a named hook to the component. Available hooks are:
-     *
-     *      * request.rendered
-     */
     register(hookName: string, callback: () => void): void {
         const hooks = this.hooks.get(hookName) || [];
         hooks.push(callback);
+        this.hooks.set(hookName, hooks);
+    }
+
+    unregister(hookName: string, callback: () => void): void {
+        const hooks = this.hooks.get(hookName) || [];
+
+        const index = hooks.indexOf(callback);
+        if (index === -1) {
+            return;
+        }
+
+        hooks.splice(index, 1);
         this.hooks.set(hookName, hooks);
     }