+++ src/TwigComponent/src/DependencyInjection/TwigComponentExtension.php
@@ -25,6 +25,7 @@ use Symfony\UX\TwigComponent\ComponentRenderer;
 use Symfony\UX\TwigComponent\ComponentRendererInterface;
 use Symfony\UX\TwigComponent\ComponentStack;
 use Symfony\UX\TwigComponent\DependencyInjection\Compiler\TwigComponentPass;
+use Symfony\UX\TwigComponent\EventListener\DataModelPropsSubscriber;
 use Symfony\UX\TwigComponent\Twig\ComponentExtension;
 
 /**
@@ -53,6 +54,7 @@ final class TwigComponentExtension extends Extension
             ->setArguments([
                 new ServiceLocatorArgument(new TaggedIteratorArgument('twig.component', 'key', null, true)),
                 new Reference('property_accessor'),
+                new Reference('event_dispatcher'),
                 class_exists(AbstractArgument::class) ? new AbstractArgument(sprintf('Added in %s.', TwigComponentPass::class)) : [],
             ])
         ;
@@ -75,6 +77,14 @@ final class TwigComponentExtension extends Extension
             ->addTag('twig.extension')
             ->addTag('container.service_subscriber', ['key' => ComponentRenderer::class, 'id' => 'ux.twig_component.component_renderer'])
             ->addTag('container.service_subscriber', ['key' => ComponentFactory::class, 'id' => 'ux.twig_component.component_factory'])
+        ;
+
+        $container->register('ux.twig_component.event_listener.data_model_props_subscriber', DataModelPropsSubscriber::class)
+            ->addTag('kernel.event_subscriber')
+            ->setArguments([
+                new Reference('ux.twig_component.component_stack'),
+                new Reference('property_accessor'),
+            ])
         ;
     }
 }